+++ docs/changelog.md
@@ -15,6 +15,7 @@ Audio driver changes:
  * Added `set_echo_feedback` and `adjust_echo_feedback` instructions
  * Added `set_fir_filter`, `set_fir_tap` and `adjust_fir_tap` instructions
  * Added limit to adjust echo feedback and adjust FIR tap instructions
+ * Added `set_channel_invert` instruction (`i` MML command)
 
 MML changes:
  * `#EchoVolume` now accepts a stereo input